Added basic structure with header, footer and home components
