+++ src/components/Footer.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+class Footer extends Component {
+
+  render () {
+    return (
+      <div> Hello footer </div>
+    );
+  }
+}
+
+export default Footer;
