+++ src/components/Home.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+class Home extends Component {
+
+  render () {
+    return (
+      <div> Hello home </div>
+    );
+  }
+}
+
+export default Home;
